# Use auto-instrumentation SDK in global tracing