--- a/.golangci.yml
+++ b/.golangci.yml
@@ -95,6 +95,17 @@ linters-settings:
           - pkg: "crypto/md5"
           - pkg: "crypto/sha1"
           - pkg: "crypto/**/pkix"
+      auto/sdk:
+        files:
+          - "!internal/global/trace.go"
+          - "~internal/global/trace_test.go"
+        deny:
+          - pkg: "go.opentelemetry.io/auto/sdk"
+            desc: Do not use SDK from automatic instrumentation.
+          - pkg: "go.opentelemetry.io/auto/sdk/telemetry"
+            desc: Do not use data-model from automatic instrumentation.
+          - pkg: "go.opentelemetry.io/auto/sdk/telemetry/test"
+            desc: Do not import data-model tests from automatic instrumentation.
       otlp-internal:
         files:
           - "!**/exporters/otlp/internal/**/*.go"