--- a/.golangci.yml
+++ b/.golangci.yml
@@ -104,10 +104,6 @@ linters-settings:
         deny:
           - pkg: "go.opentelemetry.io/auto/sdk"
             desc: Do not use SDK from automatic instrumentation.
-          - pkg: "go.opentelemetry.io/auto/sdk/telemetry"
-            desc: Do not use data-model from automatic instrumentation.
-          - pkg: "go.opentelemetry.io/auto/sdk/telemetry/test"
-            desc: Do not import data-model tests from automatic instrumentation.
       otlp-internal:
         files:
           - "!**/exporters/otlp/internal/**/*.go"