--- a/bridge/opencensus/go.mod
+++ b/bridge/opencensus/go.mod
@@ -18,7 +18,7 @@ require (
 	github.com/golang/groupcache v0.0.0-20210331224755-41bb18bfe9da // indirect
 	github.com/google/uuid v1.6.0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	go.opentelemetry.io/auto/sdk v0.1.0-alpha // indirect
+	go.opentelemetry.io/auto/sdk v0.2.0-alpha // indirect
 	go.opentelemetry.io/otel/metric v1.31.0 // indirect
 	golang.org/x/sys v0.26.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect