--- a/bridge/opentracing/go.mod
+++ b/bridge/opentracing/go.mod
@@ -18,7 +18,7 @@ require (
 	github.com/go-logr/logr v1.4.2 // indirect
 	github.com/go-logr/stdr v1.2.2 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	go.opentelemetry.io/auto/sdk v0.1.0-alpha // indirect
+	go.opentelemetry.io/auto/sdk v0.2.0-alpha // indirect
 	go.opentelemetry.io/otel/metric v1.31.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )