--- a/exporters/prometheus/go.mod
+++ b/exporters/prometheus/go.mod
@@ -1,6 +1,6 @@
 module go.opentelemetry.io/otel/exporters/prometheus
 
-go 1.22
+go 1.22.0
 
 require (
 	github.com/prometheus/client_golang v1.20.5
@@ -22,11 +22,11 @@ require (
 	github.com/go-logr/logr v1.4.2 // indirect
 	github.com/go-logr/stdr v1.2.2 // indirect
 	github.com/google/uuid v1.6.0 // indirect
-	github.com/kr/text v0.2.0 // indirect
 	github.com/kylelemons/godebug v1.1.0 // indirect
 	github.com/munnerz/goautoneg v0.0.0-20191010083416-a7dc8b61c822 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/prometheus/procfs v0.15.1 // indirect
+	go.opentelemetry.io/auto/sdk v0.1.0-alpha // indirect
 	golang.org/x/sys v0.26.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )