--- a/log/go.mod
+++ b/log/go.mod
@@ -1,6 +1,6 @@
 module go.opentelemetry.io/otel/log
 
-go 1.22
+go 1.22.0
 
 require (
 	github.com/go-logr/logr v1.4.2
@@ -12,6 +12,7 @@ require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/go-logr/stdr v1.2.2 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
+	go.opentelemetry.io/auto/sdk v0.1.0-alpha // indirect
 	go.opentelemetry.io/otel/metric v1.31.0 // indirect
 	go.opentelemetry.io/otel/trace v1.31.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect