--- a/sdk/log/go.mod
+++ b/sdk/log/go.mod
@@ -1,6 +1,6 @@
 module go.opentelemetry.io/otel/sdk/log
 
-go 1.22
+go 1.22.0
 
 require (
 	github.com/go-logr/logr v1.4.2
@@ -17,6 +17,7 @@ require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/google/uuid v1.6.0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
+	go.opentelemetry.io/auto/sdk v0.1.0-alpha // indirect
 	go.opentelemetry.io/otel/metric v1.31.0 // indirect
 	golang.org/x/sys v0.26.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect